GPGS Implemented
Added manual login / logout. Added achievements and leaderboard. Fixed
some graphical scaling issues.

+++ core/src/com/lunarcannon/NanoCat/ExternalInterface.java
@@ -9,7 +9,7 @@ public void logout();
 
 public boolean getSignedIn();
 
-public void submitScore(int highScore);
+public void submitScore(long highScore);
 
 public void getScore();
